raspberry_pi/charli-home/requirements.txt
soundfile
numpy
faster-whisper
openai
fastapi
uvicorn[standard]
websockets
pvporcupine
pvrecorder
psutil
textual
websocket-client
